Fix ghci linker on x32
Integer literals used to detect too-long relocations were 'long',
which is not long enough on ILP32. That meant every relocation was
considered too long.

The way too-long relocations are handled assumes that
1. the target of the relocation is used as code (jumped to)
2. the addend is such that the destination of the jump is the symbol
   (addend = -4 for a relative jmp/call, addend = 0 for a Haskell
    pushed return address).
Jump tables used to compile switch statements violate 1: they are
tables of code pointers, not code themselves. The relocations in
.init_array sections used to violate 2, since they were pointers to
static functions, and represented as .text+0abcde.

With the corrected logic for detecting too-long relocations,
things seem to be mostly okay: the remaining too-long relocations
are to regular exported C functions.

TODO: Add an assertion about addend on a too-long relocation,
maybe this is the cause of some plug-in segfaults.